+++ cookbook/facebook_visualization/download_and_extract.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+wget --no-check-certificate https://snap.stanford.edu/data/facebook_combined.txt.gz
+gunzip facebook_combined.txt.gz