+++ docs/conf.py
@@ -58,6 +58,7 @@ extensions = [
     'sphinx.ext.autosummary',
     'sphinx.ext.napoleon',
     'rst2pdf.pdfbuilder',
+    'sphinxcontrib.video',
 ]
 
 if USE_KATEX:
@@ -213,7 +214,8 @@ html_static_path = ['_static']
 
 # If true, SmartyPants will be used to convert quotes and dashes to
 # typographically correct entities.
-#html_use_smartypants = True
+#html_use_smartypants = False
+smartquotes = False
 
 # Custom sidebar templates, maps document names to template names.
 #html_sidebars = {}